requirements.txt
click
psutil
python-apt